feat: add error msg search family
